requirements.txt
django>=4.2,<5.0
flake8>=6.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0